Minor refactoring

+++ bq-api/src/models/tasks.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 /// Represents the structure of user tasks corresponding to their JSON representation.
-pub type UserTaskEntries = HashMap<u16, HashMap<u16, HashMap<u16, UserTask>>>;
+pub type UserTaskEntries = HashMap<usize, HashMap<usize, HashMap<usize, UserTask>>>;
 
 #[derive(Serialize, Deserialize)]
 pub struct SubTask {
